🎨 모델 파일 분리
+ MyPageItemInfo
+ MyPageItemType
